Update 1.2
- Corrected the misconception of R,I,Z  instead of |Z|,|phase|, R|Z|, I|Z|
- Currently working with the how the board overwrites previous readings if too many are stored(not yet working correctly)
- Phase angle is finally being calculated, as well as used to find the actual R/I values.
- Gain calculation is done at startup(no recalculation option yet)(this will be changed on a new hardware)
- General fixes related to the above topics.

+++ mbi-core-ui/proto_UI/arduino-ad5933-master/AD5933.h
@@ -141,8 +141,8 @@ class AD5933 {
 
         // Perform frequency sweeps
         static bool frequencySweep(int real[], int imag[], int);
-        static bool calibrate(double gain[], int phase[], int ref, int n);
-        static bool calibrate(double gain[], int phase[], int real[],
+        static bool calibrate(float gain[], double phase[], int ref, int n);
+        static bool calibrate(float gain[], double phase[], int real[],
                               int imag[], int ref, int n);
     private:
         // Private data
